refactor: remove wrappers files and update components
