+++ packages/app/src/App.native.tsx
@@ -14,11 +14,12 @@ import * as Routing from './routes/routing';
 import ActivityLogPage from './pages/ActivityLogPage';
 import { Providers } from './Providers';
 import DonatePage from './pages/DonatePage';
-import { WagmiProvider } from 'wagmi';
+import { createConfig, http, WagmiProvider } from 'wagmi';
 import { celo, mainnet, AppKitNetwork } from '@reown/appkit/networks'
+import { injected, walletConnect } from 'wagmi/connectors'
 import { ApolloProvider } from '@apollo/client';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import { createAppKit } from '@reown/appkit/react';
+import { createAppKit } from '@reown/appkit';
 import { Colors } from './utils/colors';
 import { WagmiAdapter } from '@reown/appkit-adapter-wagmi'
 import { useCreateSubgraphApolloClient, useCreateMongoDbApolloClient } from './hooks/apollo';
@@ -45,14 +46,27 @@ createAppKit({
   networks,
   projectId,
   metadata,
-  defaultNetwork: celo,
-  features: {
-   email: false,
-   socials: false,
-  }
+  defaultNetwork: celo
 })
 
-function AppWep(): JSX.Element {
+function App(): JSX.Element {
+  const connectors = [
+    injected({
+      target: 'metaMask',
+    }),
+    walletConnect({
+        projectId: 'f147afbc9ad50465eaedd3f56ad2ae87',
+    }),
+  ];
+
+  const wagmiConfig = createConfig({
+    connectors,
+    chains: [celo], 
+    transports: { 
+      [celo.id]: http("https://rpc.ankr.com/celo"), 
+    },
+  });
+
   const subgraphApolloClient = useCreateSubgraphApolloClient();
   const mongoDbApolloClient = useCreateMongoDbApolloClient();
 
@@ -62,7 +76,7 @@ function AppWep(): JSX.Element {
 
   return (
     <Providers>
-      <WagmiProvider config={wagmiAdapter.wagmiConfig}>
+      <WagmiProvider config={wagmiConfig}>
       <QueryClientProvider client={queryClient}>
         <ApolloProvider client={subgraphApolloClient}>
           <MongoDbApolloProvider client={mongoDbApolloClient}>
@@ -98,4 +112,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export default AppWep;
+export default App;
