--- packages/app/src/AppWrapper.ts
@@ -1,8 +1,0 @@
-import { Platform } from 'react-native';
-import AppWeb from './App.web';
-import AppNative from './App';
-
-export const App = Platform.select({
-  web: AppWeb,
-  default: AppNative,
-});
