--- packages/app/src/components/Header/ConnectWalletMenuWrapper.ts
@@ -1,8 +1,0 @@
-import { Platform } from 'react-native';
-import { ConnectWalletMenuWeb }  from './ConnectWalletMenu.web';
-import { ConnectWalletMenu }  from './ConnectWalletMenu';
-
-export const ConnectWalletMenuWrapper = Platform.select({
-  web: ConnectWalletMenuWeb,
-  default: ConnectWalletMenu,
-});
