+++ packages/app/src/components/Header/Header.tsx
@@ -2,9 +2,8 @@ import { Image, StyleSheet, TouchableOpacity, View } from 'react-native';
 import { useAccount } from 'wagmi';
 
 import { ConnectedAccountDisplay } from './ConnectedAccountDisplay';
-import { ConnectWalletMenuWrapper } from './ConnectWalletMenuWrapper';
+import { ConnectWalletMenu } from './ConnectWalletMenu';
 import { DropdownMenu } from './DropdownMenu';
-
 import useCrossNavigate from '../../routes/useCrossNavigate';
 import { Colors } from '../../utils/colors';
 import { BackIcon, HeaderLogo } from '../../assets';
@@ -26,7 +25,7 @@ function Header(): JSX.Element {
           </View>
           <View style={[styles.walletInfoContainer, styles.walletInfoContainerDesktop]}>
             {address && <ConnectedAccountDisplay isDesktopResolution={isDesktopView} address={address} />}
-            {!address && <ConnectWalletMenuWrapper dropdownOffset={{ top: 40, right: 50 }} />}
+            {!address && <ConnectWalletMenu dropdownOffset={{ top: 40, right: 50 }} />}
             <DropdownMenu
               isDesktopResolution={isDesktopView}
               address={address}
@@ -39,7 +38,7 @@ function Header(): JSX.Element {
         <View style={styles.headerMobileContainer}>
           <View style={[styles.walletInfoContainer, styles.desktopWrapper]}>
             {address && <ConnectedAccountDisplay isDesktopResolution={isDesktopView} address={address} />}
-            {!address && <ConnectWalletMenuWrapper dropdownOffset={{ top: 37, left: 0 }} />}
+            {!address && <ConnectWalletMenu dropdownOffset={{ top: 37, left: 0 }} />}
             <DropdownMenu
               isDesktopResolution={isDesktopView}
               address={address}
