+++ packages/app/web/main.tsx
@@ -1,7 +1,7 @@
 import './shim';
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import { App } from '../src/AppWrapper';
+import App from '../src/App';
 
 // import { configureChains, createConfig, WagmiConfig } from 'wagmi';
 // import { celo } from 'wagmi/chains';
